allow to specify a lock which will be acquired during job execution
It's now possible to specify a lock id as a job parameter. This lock will be acquired prior to running the job and released after the job finished running. This also changes the way we handle job active/inactive management which is now entirely the responsibility of the job state synchronizer.