need to handle hashCode and equals within the proxy
In order to allow the proxy to properly work in collections we must
handle equals and hashCode in a special way. One way is to only rely on
exact identity and that is what we do. While we do this we can also
improve the toString output to help identifying that this is a proxy.

Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>