replace custom test application with JUnit 4 rule/resource
This commit removes the custom Gyrex test application. Thus, instead of
patching PDE and Tycho to launch a custom app a JUnit 4 standard
mechanism has been adopted. A new rule is now available, which can be
used in suites and tests to require the start of a Gyrex server. Rules
can be nested. 

However, the drawback of this approach is that clients must be carful
what to rely on in their bundle activators. For example, test bundles
must be aware that Gyrex may not be running when their bundle activator
is started. Consequently, a test bundle my trigger bundle activation for
another bundle. Such a bundle must also be prepared that a server
environment has not been initialized yet.

Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>